[Sync] Add a test for recovering from depleted quota
This CL adds a test that properly tests depleting commit quota and
recovering from it not sooner than expected.

As this test has some flaking potential, the test comes in a separate
CL. It's not critial for the test to stick on all platforms, it would be
great if it can stick at least on a few platforms.

Bug: 1145138
Change-Id: If74c5fbf85c9a4c977f25ff42c891d70ce64000c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3557212
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Jan Krčál <jkrcal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#986882}