Add instrumentation for UKM sampling of discrete-valued media features
We want to compute the entropy leaked through querying CSS media
features. The approach taken here is to sample the queried media
features once per document. What we store are key-value pairs
where the key is the name of the media feature and the value is the
media value (not the truth evaluation of the query).

To be able to store the evaluation state of a feature per document we
use a uint64_t inside document.h where we set the bits on a per feature
basis as we sample.

In a further step we plan to add the sampling code also for range media
queries.

Design doc: go/privacy-budget:instrumenting-media-queries

Bug: 1307190
Change-Id: Ic5a7c24e48c26d5265b62cf5feae4f795b8251ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3531995
Reviewed-by: Antonio Sartori <antoniosartori@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Alexandra Reimers <areimers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#986924}