[PartitionAlloc] Clamp periodic purge interval
Thread cache periodic purge frequency is adjusted at each purge to stay
between a low and high value. However, if there is a background thread
which holds onto a large amount of cached memory *and* is sleeping, its
cached memory will not be reclaimed in-between two purges, leading to
the interval shrinking way below the minimum.

This happens for the OOBE on CrOS, as seen in the linked bug. To fix it,
clamp the purge interval.

Bisected by rsorokin@.

Bug: 1311292
Change-Id: I2b2620d584ddec464eddaba13888494fec6645a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3560230
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Benoit Lize <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#986933}