[basic-card] Migrate WPTs that use basic-card identifier
With basic card now disabled and soon to be removed, migrate web
platform tests that test the payment method identifier syntax to use
secure-payment-confirmation instead of basic-card. These tests pass
either way, this is just to refresh the test data to a current feature.

The merchant-validation test is also updated, in addition to the
similar PMI validation change, it is updated to use an example URL
payment method.

Bug: 1209835
Change-Id: I109e9063720c21d10c5c0af1f765d553eaadeb83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3556267
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Commit-Queue: Nick Burris <nburris@chromium.org>
Cr-Commit-Position: refs/heads/main@{#986998}