Privacy Sandbox Settings: Consider management as confirmation
The logic which checked to see whether a user had confirmed the Privacy
Sandbox did not account for users which had the Sandbox controlled
elsewhere, e.g. via policy.

CL adjusts the logic to consider a Privacy Sandbox preference which is
not user control-able as confirmation. Noting that a confirmed Privacy
Sandbox must also be enabled to function.

Bug: 1311644
Change-Id: I63a95408a2c4fee21e6a348c890b3cb55ea7558b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3560838
Reviewed-by: Olesia Marukhno <olesiamarukhno@google.com>
Commit-Queue: Theodore Olsauskas-Warren <sauski@google.com>
Cr-Commit-Position: refs/heads/main@{#987020}