[Android Tab Restore] Introduce TabModelObserver#didCloseTabs()
Sheriffs: If this CL causes flakes in UndoTabModelTest please disable
those test cases rather than reverting this CL. Many tests in that suite
were disabled long ago and I'm hoping can be revived without issue as
a result of these changes.

Reference doc:
https://docs.google.com/document/d/1KaXyNs8aOAnumY5RzVyNj6YEJBDX0yrv-iaYjgrmGJY/edit?usp=sharing&resourcekey=0-IdsMOCz46tsXuTUUuWHQ7w

Add a method didCloseTabs(List<Tab>) that is called before committing
tab closures. This is called just before one or more tabs will commit to
closing. This offers an observer the opportunity to act on a set of tabs
closed together by an "event" rather than only tab-by-tab via
didCloseTab().

This will be used to facilitate storing group/bulk closures to
TabRestoreService. See prototype CL:
https://chromium-review.googlesource.com/c/chromium/src/+/3489596

In order to achieve this behavior TabModelImpl was refactored. Now
most Pending Tab Closure logic lives in PendingTabClosureManager
which owns the RewoundList from before and an event queue of
TabClosureEvent objects to process.

When any tab closure cannot be undone nothing changes other than calling
didCloseAllTabs().

When a call to closeTab, closeMultipleTabs or closeAllTabs can be undone
it is added to PendingTabClosureManager as a single TabClosureEvent.
- When commitAllTabClosures is called, every closure in
  PendingTabClosureManager as TabClosureEvent groups in FIFO order.
- When commitTabClosure is called PendingTabClosureManager waits until
  all tabs that were added as a TabClosureEvent are ready to commit
  before committing the closure.
- When cancelTabClosure is called PendingTabClosureManager immediately
  undoes the tab closure requested, but will expect cancelTabClosure
  to be called for every Tab in the TabClosureEvent.

Note: it is not safe to commit some tabs and cancel other that were
added in the same TabClosureEvent. This is enforced via asserts to
ensure some degree of "atomicity" to the behavior of a TabClosureEvent
i.e. it must either be fully committed or fully undone.

Bug: 1306258
Change-Id: I34e37ff3c3ab1a229b00a41186bac83dcfae5159
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3523275
Reviewed-by: Fred Mello <fredmello@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987053}