Create new UrlParamClassificationComponentInstaller class.
This class will receive classification lists and validate them.
If the list is valid, it will be provided to the
url_param_filter::ClassificationsLoader.

The validation of classification lists occurs in VerifyInstallation, and
the results of this are outputted to a new "Navigation.UrlParamFilter.
ClassificationListValidationResult" histogram.

Change-Id: Ic1cacf585c7dd61992919dd9f9e4f278496d3f11
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3537547
Reviewed-by: Matt Reichhoff <mreichhoff@chromium.org>
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Reviewed-by: Ben Kelly <wanderview@chromium.org>
Commit-Queue: Kirubel Aklilu <kaklilu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987087}