MediaRecorder: Add UMA for codec
Prior to this CL, we did not have a UMA for which codec and hardware/
software variant is used by MediaRecorder. After this CL, we have a
UMA for it. This is a part of work to enable VBR in VA-API; the
design doc is available at
https://docs.google.com/document/d/1kofC1bNUgfJQ3DAlYuTlEkXhe86my0UrKzAXGHvnvgw/edit?usp=sharing

Bug: b:182227253
Change-Id: I7669e810329bb9deb0f4104ec3f67023e1461a5d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3556631
Reviewed-by: Miguel Casas-Sanchez <mcasas@chromium.org>
Reviewed-by: Evan Liu <evliu@google.com>
Commit-Queue: Clarissa Garvey <clarissagarvey@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987168}