[Fast Pair] Don't create Gatt connection if already paired to Chromebook
Exposes API in FastPairRepository to check the SavedDeviceRegistry for
an account key, and if the account key exists already, it implies that
the device has already been paired with this Chromebook. Uses this API
in the NotDiscoverableScanner to not create a Gatt connection to a
device if it has already been paired.

Test: verified on DUT
Change-Id: Iaf6cfc640ee5a7cdce49c2421e4bcc6444c51fa4
Fixed: b/227209345
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3557556
Reviewed-by: Shane Fitzpatrick <shanefitz@google.com>
Commit-Queue: Juliet Lévesque <julietlevesque@google.com>
Cr-Commit-Position: refs/heads/main@{#987196}