Fix a UAF issue
The root cause of the problem is: During `~CalendarView`, the
`CalendarMonthView` will be destroyed first. Then the abort event
callback will be run, and it will finally access `previous_month_`
which is the destroyed `CalendarMonthView`.

So need to clean up event_list before `RemoveAllChildViews`:

Bug: 1292308
Change-Id: Ib03b2d331c725fc8caea84637bc80472f732ec31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3561699
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Commit-Queue: Jiaming Cheng <jiamingc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987223}