Skip tests that use NamedTemporaryFile on Windows
tempfile.NamedTemporaryFile is a handy way to create a file that is
guaranteed to be deleted when the test finishes. On Linux you can even
pass the file name down to a function being tested and the test can open
the file. However this doesn't work on Windows.

There is limited value to having these tests work on Windows so the
pragmatic thing is just to skip them, rather than having them cause
cryptic errors that may halt presubmits.

This also maps 'gclient' to 'gclient.bat' on Windows to avoid file-not-
found errors when invoking it, necessary when shell=False.

Bug: 1309977
Change-Id: I042896056c7e374a2632d113ff893cff5aa06bb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3558374
Auto-Submit: Bruce Dawson <brucedawson@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987237}