[journeys] Separate Labels From Entities into its own flag and loop
Labels from Entities is not approved yet, and we may choose to launch
a version of labels without sourcing from entities.

To that end, this CL breaks that logic out under its own feature
parameter, and updates the test to verify that the flag turns on and
off entity-sourced labels.

Bug: 1294348
Change-Id: I1d49785fad19d0d5eda73a7bcc55415d72aecd1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3557906
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987255}