EyeDropper: Use abort reason
This change updates the open() method and abort steps attached to the
provided AbortSginal so that the reason provided to the AbortController
is used to reject the Promise returned by open().

The web-visible effect of this change is minimal because by default the
abort reason is set to the same "AbortError" DOMException which was
already used.

The corresponding specification change is:
https://github.com/WICG/eyedropper-api/pull/25

Change-Id: I414bcd2ae7cd8fba3031c66ba98a5d04b0a7c740
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3559669
Reviewed-by: Ionel Popescu <iopopesc@microsoft.com>
Commit-Queue: Nidhi Jaju <nidhijaju@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987267}