Use event source-id to filter out remote CRD input
During a CRD session input can come from 2 sources:
   - The remote user (through CRD).
   - The local user.

All remote input is blocked for 2 seconds when local input is detected,
because we do not want the remote user to steal control of the local user.

Previously this was detected by remembering the location of all remote input,
and then monitoring all local input and filtering out the echos of the
remote input.
However this is not 100% accurate, so in this CL I introduce a system
to use the event's source-id to filter out the echos of the remote
commands.

Bug: b/208370410
Test: manually deployed and tested
Change-Id: I1dfd2afe0ebcd611219509a64b9f7f7bc3faac2a
Cq-Include-Trybots: luci.chrome.try:linux-chromeos-chrome
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3440148
Reviewed-by: Ben Franz <bfranz@chromium.org>
Reviewed-by: Joe Downing <joedow@chromium.org>
Auto-Submit: Jeroen Dhollander <jeroendh@google.com>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Commit-Queue: Sadrul Chowdhury <sadrul@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987305}