Enable u2fd.WebauthnUsingPassword.tpm1
This test is fixed and we saw it continuously passing on tastboard
again. There should be no timeout issue anymore as the implementation
won't parse chrome logs anymore, and I verified that the time needed for
the test is very low and quite consistent on stainless.

Bug: 1309146
Change-Id: I66b5c10707fdf9a935733e5a3904eb3a3dd2e26f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3555494
Reviewed-by: Jiaming Cheng <jiamingc@chromium.org>
Commit-Queue: Howard Yang <hcyang@google.com>
Cr-Commit-Position: refs/heads/main@{#987310}