Fix the UninstallDuringLastBrowserWindow test for the non mojom AppService.
This CL is used to prepare for CL:3547984 to enable the
AppServiceOnAppUpdateWithoutMojom flag.

The UninstallDuringLastBrowserWindow browser test verifies the web app
browser window can be closed when uninstall the web app. With the mojom
AppService, the running process is:
install_finalizer().UninstallWebApp
  --> uninstall_task->Start [1]
    --> NotifyWebAppWillBeUninstalled [2]
      --> WebAppPublisherHelper::OnWebAppWillBeUninstalled
        --> Publish app ASYNC
          --> ShelfAppServiceAppUpdater::OnAppUpdate
            --> Close the browser window [4]
    --> BindLambdaForTesting AYNC [3]

[1] https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/web_applications/web_app_install_finalizer.cc;l=492;drc=42e75824670baadf21ece40f71b946500af1cbce
[2] https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/web_applications/web_app_uninstall_job.cc;drc=42e75824670baadf21ece40f71b946500af1cbce;l=68
[3] https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/web_applications/web_app_install_finalizer.cc;l=497;drc=42e75824670baadf21ece40f71b946500af1cbce
[4] https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/ui/ash/shelf/app_service/shelf_app_service_app_updater.cc;l=53

With the mojom AppService, the test works because mojom web apps is
is published ASYNC, so after run_loop is quit in BindLambdaForTesting,
the second removed_run_loop_ in BrowserWaiter(app_browser).AwaitRemoved
can wait for the mojom App to be updated and close the browser window.

However, with the non mojom AppService, the web app is published SYNC,
so the browser window is closed SYNC before the first run_loop quit in
BindLambdaForTesting, so the second removed_run_loop_ in
BrowserWaiter(app_browser) is never stoped because the browser window
has been closed before BrowserWaiter(app_browser) is created.

This CL modifies the test, create BrowserWaiter(app_browser) before
running the first run_loop and uninstall web apps, so that BrowserWaiter
can observe the browser list, and get the removed browser info in
BrowserWaiter::OnBrowserRemoved. If the browser has been removed, the
removed_run_loop_ doesn't need to be run again.

This CL is pure refactor, and no function change.

BUG=1253250

Change-Id: I9e22e7ca9c3a42583f69e9ce1b179249b2fbfd1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3557872
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987313}