Make Starter a WebContentUserData
Before this change, the starter was owned by the Android platform
delegate. In order to support other platforms, the starter is
now a WebContentUserData and will be created on tab creation.
This CL adds the creation for android only, a follow-up will
add the same for other platforms.

Bug: b:201964911


Change-Id: I7fea6d3d47ea09502530f25daca891e0c09ab86c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3528555
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Commit-Queue: Luca Hunkeler <hluca@google.com>
Cr-Commit-Position: refs/heads/main@{#987368}