Split ElementFinder
To make parallel execution between semantic and css
element finding possible, this splits the element finder
into two pieces, one for each.

The aim is to separate code and prevent accidental re-use
of member variables that are only valid in one of the
cases. This change fixes a bug that was found though this
(setting the frame in the semantic case).

Bug: b:224745206
Change-Id: Ia44cb7d51c2d8c2cb79e1f577605c739756f581d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3560049
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Commit-Queue: Sandro Maggi <sandromaggi@google.com>
Cr-Commit-Position: refs/heads/main@{#987386}