dlp: Always send snapshot_source_id over DLP crosapi.
As the version of the DLP interface doesn't increase when
a struct is being extended with a new field - it's not possible
to check receiver version on whether it supports the new field.
However, it should be safe to send the struct with additional field
as older receivers will just ignore it.

Bug: 1293023
Test: Try to start window share from Lacros for a confidential window.
Change-Id: Ia74a5a1c30eacf067eb08c0300802f7e0b4c792e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3557992
Reviewed-by: Aya Elsayed <ayaelattar@chromium.org>
Commit-Queue: Sergey Poromov <poromov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987398}