[Private Network Access] Write tests for same-origin PNA preflights.
There is a bug in the error ignoring logic used when PNA preflights are
sent experimentally, in warning-only mode. In the case of same-origin
non-"simple" requests, CORS skips sending a preflight but PNA logic
incorrectly assumed that a preflight would have been sent because the
request was not "simple". As a result, PNA decided that non-PNA errors
with the preflight should not be ignored. This resulted in broken
requests when preflights were experimentally enabled.

Bug: chromium:1299749
Change-Id: Ie4565a3ac184992a09ad98ad27f488aaf5648d86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3557629
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Auto-Submit: Titouan Rigoudy <titouan@chromium.org>
Commit-Queue: Titouan Rigoudy <titouan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987449}