Allow OOPR-Canvas to use hardware overlay
This change updates how the RasterDecoder initializes its GPU
capabilities in order to set the texture_storage_image bit according
to whether the image factory supports anonymous images.  This is
equivalent to how that bit gets set in GLES2 decoders.  We are using
this capability as a proxy for determining whether SharedImage
supports scanout, which is conservative.  In the long term we should
probably add a dedicated capability bit that is specific to
SharedImage functionality. This is a rapid fix for unblocking the
launch of OOPR-Canvas in M101.

Tests landing separately. See: crbug.com/1309206

Bug: 1305679
Change-Id: I854f37a0ebb76026a4f278d4ca72572372ca08cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3557831
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Justin Novosad <junov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987470}