[ozone] Fix lacros and linux power save blockers to allow stacking.
The lacros and linux implementations of `PowerSaveBlocker` have
incorrect behavior when the `ENABLE_IDLE_INHIBIT` feature flag is
enabled. When it is enabled, they currently use
`display::Screen::SetScreenSaverSuspended` to apply the block, which is
implemented in a manner that ignores stacking calls. That is to say, the
last suspension or un-suspension will have an overriding effect on the
system [1].

That approach is not compatible with current usage patterns, where
multiple blocks are sometimes created and removed at independent times.
That stacked usage is correctly handled by the ash implementation of
`PowerSaveBlocker` as well as the lacros implementation when the
`ENABLE_IDLE_INHIBIT` feature flag is disabled.

This change modifies the ozone screen saver suspension implementation
such that each suspension invocation is counted and returns a suspender
object, and the system will not un-suspend until the last suspender is
destructed. That modification corrects the lacros and linux
`PowerSaveBlocker` implementations to match with expected behaviors.

[1] https://crsrc.org/c/ui/ozone/public/platform_screen.h;l=89

Bug: b:196213351
Bug: b:193670013
Bug: 1295414
Change-Id: Iaf57161f94b85090207c24f9cb4e24b602356bb2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3507472
Reviewed-by: Brian Ho <hob@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Andrew Wolfers <aswolfers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987535}