[fencedframes] Compute frozen size for MPArch version
This patch enables the logic to compute frozen size for
the Shadow DOM, implemented at r946679 crrev.com/c/3282963,
for the MPArch.

How to compute the frozen size is still under discussion, but
the logic should be shared between the Shadow DOM and the
MPArch. In this patch, both versions use `ResizeObserver`.

Bug: 1270648
Change-Id: Ieb0be759f2ae69fc84379a6603722b026aeb9dab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3563598
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987536}