AW UMA: link to detailed doc for package names
No change to current behavior or privacy requirements. This CL is
privacy-neutral and only aims to link to an internal document to provide
more context and clear up some ambiguity:

* "Privacy team" was too vague: the document clarifies exactly which
  internal privacy team has the appropriate context to review package
  name logging in WebView UMA.
* This aims to make the comment a little briefer: some of the additional
  clarification has been moved to the internal doc (and expanded upon
  for clarity). The actual rationale for this behavior has not changed.
* The internal doc also includes other internal links so that readers
  can learn more about UMA and related subject material.

This also adds a reference to the public documentation around metrics
collection in WebView so that folks outside Google can get a better
understanding.

Test: N/A
Change-Id: Ib69868bfcd5ed7083aa708edb35ec4d329e98fae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3482077
Reviewed-by: Hazem Ashmawy <hazems@chromium.org>
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987558}