[fuchsia] Remove //fuchsia/BUILD.gn and its gn_all
This target was previously needed when `all` could not be built on
Fuchsia but is no longer needed now that the bots are building all.

Replace the dependency on //fuchsia:gn_all with dependencies on specific
production targets, which will ensure that targets in those directories
and ones they depend on (e.g., //fuchsia/base) are also processed.

Bug: 1081525
Test: autoninja -C out/fuchsia/ fuchsia/base:cr_fuchsia_base_unittests chromecast/bindings:bindings_manager_fuchsia fuchsia/engine:web_engine fuchsia/engine:web_engine_browsertests fuchsia/engine:web_engine_shell fuchsia/engine:web_engine_unittests fuchsia/mojom:fuchsia_mojo_unittests fuchsia/runners:cast_runner fuchsia/runners:cast_runner_browsertests fuchsia/runners:cast_runner_integration_tests fuchsia/runners:cast_runner_unittests fuchsia/runners:web_runner headless:headless_non_renderer media/cdm/library_cdm/clear_key_cdm:clear_key_cdm
Change-Id: I42a047da0c99cd55203baaa403b64c72b128cadc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3559553
Reviewed-by: Wez <wez@chromium.org>
Auto-Submit: David Dorwin <ddorwin@chromium.org>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Commit-Queue: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987576}